refactor: add tests for focusing function
